Review comment corrections. Decreased log printing if connected to an older version.

+++ hazelcast-client/src/main/java/com/hazelcast/client/spi/impl/ClientExecutionServiceImpl.java
@@ -109,7 +109,7 @@ public final class ClientExecutionServiceImpl implements ClientExecutionService,
     }
 
     @Probe (level = MANDATORY)
-    public long getUserExecutorQueueSize() {
+    public int getUserExecutorQueueSize() {
         return ((ThreadPoolExecutor) userExecutor).getQueue().size();
     }
